requirements.txt
flask==1.1.2
MarkupSafe==1.1.1
itsdangerous==2.0.1
jinja2==2.11.3
torch==1.7.1
torchvision==0.8.2
opencv-python-headless==4.5.3.56
Pillow==8.0.1
pandas==1.2.1
protobuf==3.20.1
werkzeug==2.0.3
matplotlib
fastai==1.0.61
imutils==0.5.4 
bitarray
tqdm
seaborn==0.12.0
boto3